+++ src/components/Notifications/NotificationItem.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { SyntheticEvent } from "react";
 
 import { createClassName } from "src/utils/classNames"
 import { Element } from "../Element/Element";
@@ -6,15 +6,23 @@ import { Element } from "../Element/Element";
 import { NotificationsItemStyled } from "./Notifications.styled"
 import { NotificationItemProps, NotificationItemElementType } from "./constants";
 
-export const NotificationItem = ({type, children, isDismissible, className, ...props}: NotificationItemProps) => {
-    const classNames = [ className, "ff-notification" ]
+export const NotificationItem = ({type, children, isDismissible, className, onCloseButtonClick, ...props}: NotificationItemProps) => {
+    const onDismissClick = (event: SyntheticEvent<HTMLDivElement>) => {
+        event.preventDefault();
+
+        if (onCloseButtonClick) {
+            onCloseButtonClick();
+        }
+    };
+
+    const classNames = [ className ];
 
     if (type) {
-        classNames.push(type)
+        classNames.push(type);
     }
 
     if (isDismissible) {
-        classNames.push("dismissible")
+        classNames.push("dismissible");
     }
 
     return (
@@ -28,7 +36,13 @@ export const NotificationItem = ({type, children, isDismissible, className, ...p
             </div>
 
             {isDismissible && (
-                <div className="dismiss-button" />
+                <div
+                    className="dismiss-button"
+                    onClick={onDismissClick}
+                    onKeyDown={onDismissClick}
+                    role="button"
+                    tabIndex={-1}
+                />
             )}
         </Element>
     );