+++ CHANGELOG.md
@@ -1,5 +1,11 @@
+# v0.21.1
+- Add `onCloseButtonClick` prop in `NotificationItem`.
+
+# v0.21.0
+- Code cleanup.
+
 # v0.20.2
-- Removed `gh-pages` package.
+- Remove `gh-pages` package.
 
 # v0.20.1
-- Initial release
+- Initial release.